Wow. Works! 16bit pngs
For a better visual, maybe we increase the convolution in vert to 1.5*stride to get some overlap,
also, we should maybe let the k0 filter be the alpha channel for the other filters and keep them at only 3 channels per + alpha=k0.
Thats saving data redundantly, so lets rather just use python script to read all kernels and do the prep for plotting pretty with k0 as alphas.
but Wow, this works... ok so now Im starting to really like OpenCV